+++ src/templates.js
@@ -606,6 +606,21 @@ function login() {
   `;
 }
 
+// logout panel
+function logout() {
+  return `
+  <span class="keyword-color">if</span>
+  <span class="bracket-color">(</span>
+  <span class="variable-color">session</span>
+  <span class="bracket-color">) {</span>
+  </br>
+  <span>&nbsp; &nbsp;<a class="function-color" onclick="handleMessageSending('logout', 'logout')">logout</a></span><span class="bracket-color">(</span>
+  <span class="bracket-color">)</span><span class="variable-color">;</span>
+  </br>
+  <span class="bracket-color">}</span>
+  `;
+}
+
 // essential panel to close html
 function tail() {
   return `
@@ -629,5 +644,6 @@ module.exports = {
   articleDetails,
   comment,
   login,
+  logout,
   tail
 };